consolidate the server files
