+++ package.json
@@ -21,9 +21,9 @@
   },
   "scripts": {
     "tsc": "tsc",
-    "prod": "NODE_ENV=production concurrently --kill-others \"npm run watch\" \"ts-node-dev --respawn --transpileOnly server-prod.ts\"",
+    "prod": "NODE_ENV=production concurrently --kill-others \"npm run watch\" \"npm run start-server\"",
     "start": "webpack-dev-server --mode development --open",
-    "start-server": "ts-node-dev --respawn --transpileOnly server-dev.ts",
+    "start-server": "ts-node-dev --respawn --transpileOnly server.ts",
     "watch": "webpack --watch"
   },
   "devDependencies": {
